# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  common.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  dto.LoginRequest:
    properties:
      captchaToken:
        description: |-
          CaptchaToken /captcha/verify 校验通过后返回的一次性凭证。
          必填：留空意味着攻击者可直接 POST /auth/login 绕过人机校验。
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - captchaToken
    - password
    - username
    type: object
  dto.RefreshTokenRequest:
    properties:
      refreshToken:
        description: |-
          RefreshToken 待刷新的 refresh token。

          **刻意不是 `binding:"required"`**（P3-B2 起）：浏览器路径下它由 HttpOnly
          cookie 携带，而前端 JS 读不到 HttpOnly cookie，也就**不可能**把它放进请求体 ——
          请求体是空的。真正「有没有凭据」的判定在 Controller 里用
          authcookie.RefreshFromRequest 统一做（请求体优先、其次 cookie，
          两者都取不到才拒绝）。

          保留该字段是为了不破坏 Swagger / curl / 第三方这类显式传参的调用方。
          若哪天有人"顺手"把 required 加回来，现象是「浏览器端 token 一过期就被
          踢到登录页」—— 与 B4 要修的那个缺陷一模一样，极难怀疑到参数绑定这一层。
        type: string
    type: object
...
  model.CaptchaVerifyRequest:
    properties:
      points:
        items:
          $ref: '#/definitions/model.Point'
        type: array
      token:
        type: string
    required:
    - points
    - token
    type: object
...
  model.Point:
    properties:
      x:
        type: integer
      "y":
        type: integer
    type: object
  vo.LoginResponse:
    properties:
      accessToken:
        type: string
      expiresIn:
        type: integer
      refreshToken:
        type: string
      tokenType:
        type: string
    type: object
  vo.MenuInfo:
    properties:
      children:
        items:
          $ref: '#/definitions/vo.MenuInfo'
        type: array
      component:
        type: string
      icon:
        type: string
      id:
        type: integer
      isCache:
        type: integer
      name:
        type: string
      parentId:
        type: integer
      path:
        type: string
      redirect:
        type: string
      sort:
        type: integer
      title:
        type: string
      type:
        type: integer
      visible:
        type: integer
    type: object
  vo.RoleInfo:
    properties:
      code:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  vo.UserInfoResponse:
    properties:
      avatar:
        type: string
      buttons:
        items:
          type: string
        type: array
      email:
        type: string
      id:
        type: integer
      menus:
        items:
          $ref: '#/definitions/vo.MenuInfo'
        type: array
      nickname:
        type: string
      phone:
        type: string
      roles:
        items:
          $ref: '#/definitions/vo.RoleInfo'
        type: array
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Gin-Admin 后台管理系统 API 文档
  title: Gin-Admin API
  version: 1.0.0
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: 登录参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/vo.LoginResponse'
              type: object
      summary: 用户登录
      tags:
      - 认证
  /auth/logout:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 退出登录
      tags:
      - 认证
  /auth/refresh:
    post:
      consumes:
      - application/json
      parameters:
      - description: RefreshToken（浏览器路径下可省略，凭据在 HttpOnly cookie 里）
        in: body
        name: body
        schema:
          $ref: '#/definitions/dto.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/vo.LoginResponse'
              type: object
      summary: 刷新Token
      tags:
      - 认证
  /auth/userInfo:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/vo.UserInfoResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 获取用户信息
      tags:
      - 认证
...
  /captcha/verify:
    post:
      consumes:
      - application/json
      parameters:
      - description: 验证码 token 与点击坐标
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.CaptchaVerifyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      summary: 校验点选验证码
      tags:
      - 验证码
...
  /system/pay/order:
    get:
      parameters:
      - description: 订单号
        in: query
        name: orderNo
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 按订单号查订单
      tags:
      - 支付订单
    post:
      consumes:
      - application/json
      parameters:
      - description: 订单标题
        in: body
        name: subject
        required: true
        schema:
          type: string
      - description: 订单描述
        in: body
        name: body_
        schema:
          type: string
      - description: 金额（单位：分）
        in: body
        name: amount
        required: true
        schema:
          type: integer
      - description: 支付渠道：wechat / alipay
        in: body
        name: channel
        required: true
        schema:
          type: string
      - description: 微信 openid（JSAPI 支付需要）
        in: body
        name: openId
        schema:
          type: string
      - description: 附加数据，原样回传
        in: body
        name: extra
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 创建支付订单
      tags:
      - 支付订单
  /system/pay/order/close:
    post:
      consumes:
      - application/json
      description: 仅未支付的订单可关闭；状态冲突返回 400
      parameters:
      - description: 订单号
        in: body
        name: orderNo
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 关闭订单
      tags:
      - 支付订单
  /system/pay/order/list:
    get:
      parameters:
      - description: 订单标题（模糊）
        in: query
        name: subject
        type: string
      - description: 支付渠道
        in: query
        name: channel
        type: string
      - description: 订单状态（非法值按不过滤处理）
        in: query
        name: status
        type: integer
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页条数
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 支付订单列表
      tags:
      - 支付订单
  /system/pay/order/query:
    get:
      description: 只回 orderNo / status / paidAt 三个字段，供前端轮询支付结果
      parameters:
      - description: 订单号
        in: query
        name: orderNo
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 查询订单状态（轻量）
      tags:
      - 支付订单
  /system/pay/order/refund:
    post:
      consumes:
      - application/json
      parameters:
      - description: 订单号
        in: body
        name: orderNo
        required: true
        schema:
          type: string
      - description: 退款金额（单位：分）
        in: body
        name: refundAmt
        required: true
        schema:
          type: integer
      - description: 退款单号，不传则自动生成
        in: body
        name: refundNo
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 订单退款
      tags:
      - 支付订单
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: '输入格式: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"